# openapi.yaml
openapi: 3.0.3
info:
  title: ParkVech API
  version: 1.0.0
  description: |
    OpenAPI (Swagger) spec for the ParkVech backend. Generated from your
    Express route files. Endpoints that require authentication are marked
    with `security: - bearerAuth: []` (JWT bearer token expected in the
    Authorization header).
servers:
  - url: http://localhost:5000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Booking:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        parkingSpaceId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        paymentStatus:
          type: string
          example: pending
...
    KYCSubmission:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        status:
          type: string
          example: pending
...
security:
  - bearerAuth: []
...
paths:
...
  /api/booking:
    post:
      tags:
        - booking
      summary: Create a booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Booking created
...
  /api/kyc:
    post:
      tags:
        - kyc
      summary: Submit KYC (multipart/form-data)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                idDocument:
                  type: string
                  format: binary
                otherFields:
                  type: string
      responses:
        '201':
          description: KYC submission received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCSubmission'
...
  /api/kyc/admin:
    get:
      tags:
        - kyc
      summary: Admin - list all KYC submissions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of KYC submissions
...
  /api/parking:
    post:
      tags:
        - parking
      summary: Register a new parking space
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                location:
                  type: object
      responses:
        '201':
          description: Parking space created
    get:
      tags:
        - parking
      summary: Get list of parking spaces
      parameters:
        - in: query
          name: lat
          schema:
            type: number
        - in: query
          name: lng
          schema:
            type: number
      responses:
        '200':
          description: Array of parking spaces
...